refactor : Battle Data class과 calculation 함수를 공통적으로 활용할 수 있도록 Battle 네이밍 제거 #PAR-340
