+++ core/model/src/main/java/online/partyrun/partyrunapplication/core/model/running_result/ui/RunnerRecordUiModel.kt
@@ -2,7 +2,7 @@ package online.partyrun.partyrunapplication.core.model.running_result.ui
 
 import java.time.LocalDateTime
 
-data class BattleRunnerRecordUiModel(
+data class RunnerRecordUiModel(
     val altitude: Double,
     val latitude: Double,
     val longitude: Double,
